Move generic rpc_client functions from wallet/ to client/
